Move the mock_old_api fixture to conftest.py so it can be shared
